+++ gdsfactory/cross_section.py
@@ -34,7 +34,7 @@ class CrossSection(BaseModel):
     info: Dict[str, Any] = {}
     aliases: Dict[str, Section] = {}
     name: Optional[str] = None
-    add_pins: Optional[Callable] = None
+    decorator: Optional[Callable] = None
 
     def add(
         self,
@@ -183,7 +183,7 @@ def cross_section(
     start_straight_length: float = 10e-3,
     end_straight_length: float = 10e-3,
     snap_to_grid: Optional[float] = None,
-    add_pins: Optional[Callable] = None,
+    decorator: Optional[Callable] = None,
 ) -> CrossSection:
     """Return CrossSection.
 
@@ -205,7 +205,7 @@ def cross_section(
         start_straight_length: for routing
         end_straight_length: for routing
         snap_to_grid: can snap points to grid when extruding the path
-        add_pins: adds pins to each port.
+        decorator: optional decorator function (adds pins to each port, etc.)
     """
     info = dict(
         width=width,
@@ -226,7 +226,7 @@ def cross_section(
         port_names=port_names,
     )
 
-    x = CrossSection(info=info, add_pins=add_pins)
+    x = CrossSection(info=info, decorator=decorator)
     x.add(
         width=width,
         offset=0,
@@ -307,18 +307,18 @@ def pin(
 
     .. code::
 
-                                         layer
-                                 |<------width------>|
-                                  ____________________ contact_gap              slab_gap
-                                 |                   |<----------->|             <-->
-        ___ _____________________|                   |__________________________|___
-       |   |         |                 undoped Si                  |            |   |
-       |   |    P++  |                 intrinsic region            |     N++    |   |
-       |___|_________|_____________________________________________|____________|___|
-                                                                    <----------->
-                                                                    contact_width
-       <---------------------------------------------------------------------------->
-                                      slab_width
+                                      layer
+                                |<----width--->|
+                                 _______________ contact_gap           slab_gap
+                                |              |<----------->|             <-->
+        ___ ____________________|              |__________________________|___
+       |   |         |                                       |            |   |
+       |   |    P++  |         undoped silicon               |     N++    |   |
+       |___|_________|_______________________________________|____________|___|
+                                                              <----------->
+                                                              contact_width
+       <---------------------------------------------------------------------->
+                                   slab_width
     """
     slab_width = width + 2 * contact_gap + 2 * contact_width - 2 * slab_gap
     contact_offset = width / 2 + contact_gap + contact_width / 2
@@ -407,7 +407,7 @@ def pn(
         gap_low_doping: from waveguide center to low doping
         gap_medium_doping: from waveguide center to medium doping.
             None removes medium doping
-        gap_high_doping: from waveguide center to high doping. None removes high doping
+        gap_high_doping: from center to high doping. None removes it.
         width_doping:
         width_slab:
         layer_p:
@@ -909,7 +909,7 @@ cross_section_factory = dict(
 
 if __name__ == "__main__":
     import gdsfactory as gf
-    from gdsfactory.add_pins import add_pins_siepic
+    from gdsfactory.add_pins import add_pins_siepic_optical
 
     P = gf.path.straight()
     # P = gf.path.euler(radius=10, use_eff=True)
@@ -940,7 +940,7 @@ if __name__ == "__main__":
     # X.add(width=2.0, offset=-4, layer=LAYER.HEATER, ports=["e1", "e2"])
     # X.add(width=0.5, offset=0, layer=LAYER.SLAB90, ports=["o1", "o2"])
 
-    X = rib_heater_doped(with_bot_heater=False, add_pins=add_pins_siepic)
+    X = rib_heater_doped(with_bot_heater=False, decorator=add_pins_siepic_optical)
     P = gf.path.straight(npoints=100, length=10)
 
     c = gf.path.extrude(P, X)